treewide: adapt drivers to support split muic
We use the macro IS_LEGACY on s2mu106 drivers to tell the compiler that the newer muic headers shouldn't be included but, instead, the old driver's ones. Also, some function declarations have to be excluded too as they're not available in the muic_old driver.

The header paths are also updated to the muic_old path where needed.

Signed-off-by: Nahuel Gómez <nahuelgomez329@gmail.com>