Revert "firmware_loader: Block path traversal"
It breaks the scsc drivers (a.k.a wlbt / Wi-Fi and Bluetooth drivers) used by our devices.
The reason is that the driver tries to load the firmware from "../etc/wifi/mx140.bin", but the firmware loader denies it.

[  110.969514] wlbt: __mx140_file_get_fw: firmware_variant=mx140 ()
[  110.969537] wlbt: __mx140_file_get_fw: Get WLBT fw ../etc/wifi/mx140.bin
[  110.969563] scsc_wlbt 11a70000.scsc_wifibt: Firmware load for '../etc/wifi/mx140.bin' refused, path contains '..' component
[  110.969574] wlbt: __mx140_file_get_fw: Error Loading FW, error -22
[  110.969583] wlbt: mxman_res_fw_init: mx140_file_get_fw() failed (-22)
[  110.969592] wlbt: mxman_start_boot: Error mxman_res_fw_init

This could be fixed in the driver itself but, I don't really care.

This reverts commit c30558e6c5c9ad6c86459d9acce1520ceeab9ea6.

Signed-off-by: Nahuel Gómez <nahuelgomez329@gmail.com>