input: nt36523_spi: don't export any sysfs nodes at all
These cause a kernel panic due to the new version of sec_input already exporting most of these.
It seems like we don't need these, and a look at the same driver found in the a54x kernel source shows that the sec_fn part of the driver isn't even built.

Signed-off-by: Nahuel Gómez <nahuelgomez329@gmail.com>